Remove unnecessary input field on photo
Although Sequel luckily doesn't instantiate the nested photo object of
"_delete" is passed in, ActiveRecord does. And what happens is that the
photo can get instantiated with the cached attachment before destroying
(if the user didn't refresh the page since the attachment was promoted),
which means that the stored attachment wouldn't get deleted.

We don't need this field so we can safely remove it.